User only able to delete if they own the item.

+++ backend/src/resolvers/Mutation.js
@@ -1,14 +1,14 @@
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const { randomBytes } = require('crypto');
-const { promisify } = require('util');
-const { transport, makeANiceEmail } = require('../mail');
-const { hasPermission } = require('../utils' );
+const bcrypt = require( 'bcryptjs' );
+const jwt = require( 'jsonwebtoken' );
+const { randomBytes } = require( 'crypto' );
+const { promisify } = require( 'util' );
+const { transport, makeANiceEmail } = require( '../mail' );
+const { hasPermission } = require( '../utils' );
 
 const Mutations = {
-	async createItem(parent, args, ctx, info) {
-		if(!ctx.request.userId) {
-			throw new Error('You must be logged in to add or edit items.');
+	async createItem( parent, args, ctx, info ) {
+		if( !ctx.request.userId ) {
+			throw new Error( 'You must be logged in to add or edit items.' );
 		}
 
 		const item = await ctx.db.mutation.createItem(
@@ -26,10 +26,9 @@ const Mutations = {
 			},
 			info
 		);
-		console.log(item);
 		return item;
 	},
-	updateItem(parent, args, ctx, info) {
+	updateItem( parent, args, ctx, info ) {
 		// first take a copy of the updates
 		const updates = { ...args };
 		// remove the ID from the updates
@@ -45,20 +44,25 @@ const Mutations = {
 			info
 		);
 	},
-	async deleteItem(parent, args, ctx, info) {
+	async deleteItem( parent, args, ctx, info ) {
 		const where = { id: args.id };
 		// 1. Find the item
-		const item = await ctx.db.query.item({ where }, `{ id title}`);
+		const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
 		// 2. Check if they own that item, or have the permissions.
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some( permission => ['ADMIN', 'ITEMDELETE'].includes( permission ));
 
+		if ( !ownsItem && hasPermissions ) {
+			throw new Error( 'Do not have permission to Delete Items.' );
+		}
 		// 3. Delete item.
 		return ctx.db.mutation.deleteItem({ where }, info);
 	},
-	async signup(parent, args, ctx, info) {
+	async signup( parent, args, ctx, info ) {
 		// lowercase their email
 		args.email = args.email.toLowerCase();
 		// hash their password
-		const password = await bcrypt.hash(args.password, 10);
+		const password = await bcrypt.hash( args.password, 10 );
 		// create the user in the database
 		const user = await ctx.db.mutation.createUser(
 			{
@@ -71,33 +75,33 @@ const Mutations = {
 			info
 		);
 		// create the JWT token for them
-		const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
+		const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET );
 		// Set the jwt as a cookie on the response
-		ctx.response.cookie('token', token, {
+		ctx.response.cookie( 'token', token, {
 			httpOnly: true,
 			maxAge: 1000 * 60 * 60 * 24 * 365 // 1 year cookie
 		});
 		// Return the User to the browser
 		return user;
 	},
-	async signin(parent, { email, password }, ctx, info) {
+	async signin(parent, { email, password }, ctx, info ) {
 		// 1. check if there is a user with that email
-		const user = await ctx.db.query.user({ where: { email } });
-		if (!user) {
-			throw new Error(`No such user found for email ${email}`);
+		const user = await ctx.db.query.user({ where: { email } } );
+		if ( !user ) {
+			throw new Error( `No such user found for email ${email}` );
 		}
 
 		// 2. Check if their password is correct
-		const valid = await bcrypt.compare(password, user.password);
-		if (!valid) {
-			throw new Error('Invalid Password!');
+		const valid = await bcrypt.compare( password, user.password );
+		if ( !valid ) {
+			throw new Error( 'Invalid Password!' );
 		}
 
 		// 3. Generate the JWT Token
 		const token = jwt.sign({ userId: user.id }, process.env.APP_SECRET);
 
 		// 4. Set the cookie with the token
-		ctx.response.cookie('token', token, {
+		ctx.response.cookie( 'token', token, {
 			httpOnly: true,
 			maxAge: 1000 * 60 * 60 * 24 * 365
 		});
@@ -105,21 +109,21 @@ const Mutations = {
 		return user;
 	},
 
-	signout(parent, args, ctx, info) {
+	signout( parent, args, ctx, info ) {
 		ctx.response.clearCookie('token');
 		return { message: 'Goodbye!' };
 	},
 
-	async requestReset(parent, args, ctx, info) {
+	async requestReset( parent, args, ctx, info ) {
 		// 1. Check if this is a real user
 		const user = await ctx.db.query.user({ where: { email: args.email } });
 
-		if (!user) {
-			throw new Error(`No such user found for email ${email}`);
+		if ( !user ) {
+			throw new Error(`No such user found for email ${ email }` );
 		}
 		// 2. Set a reset token and expiry on that user
-		const randomBytesPromisified = promisify(randomBytes);
-		const resetToken = (await randomBytesPromisified(20)).toString('hex');
+		const randomBytesPromisified = promisify( randomBytes );
+		const resetToken = ( await randomBytesPromisified( 20 )).toString( 'hex' );
 		const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now
 		const res = await ctx.db.mutation.updateUser({
 			where: { email: args.email },
@@ -140,9 +144,9 @@ const Mutations = {
 		// 4. Return the message
 		return { message: 'thanks!' };
 	},
-	async resetPassword(parent, args, ctx, info) {
+	async resetPassword( parent, args, ctx, info ) {
 		// 1.  check if the passwords match
-		if (args.password !== args.confirmPassword) {
+		if ( args.password !== args.confirmPassword ) {
 			throw new Error("Your Passwords don't match!");
 		}
 
@@ -155,10 +159,10 @@ const Mutations = {
 			}
 		});
 		if (!user) {
-			throw new Error('This token is either invalid or expired!');
+			throw new Error( 'This token is either invalid or expired!' );
 		}
 		// 4. Hash their new password
-		const password = await bcrypt.hash(args.password, 10);
+		const password = await bcrypt.hash(args.password, 10 );
 		// 5. Save the new password to the user and remove old resetToken fields
 		const updatedUser = await ctx.db.mutation.updateUser({
 			where: { email: user.email },
